+++ example/lib/main.dart
@@ -13,26 +13,31 @@ void main() {
 class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
 
+  final Widget drawerHeader = const Text(
+    'Hello Flutter',
+    style: TextStyle(
+      color: Colors.white,
+      fontSize: 24,
+      fontWeight: FontWeight.bold,
+      backgroundColor: Colors.blue,
+    ),
+    textAlign: TextAlign.center,
+  );
+
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider(
       create: (_) => MyCounter(),
-      child: const MaterialApp(
+      child: MaterialApp(
         debugShowCheckedModeBanner: false,
         home: BoxDrawer(
-          drawer: MyDrawer(),
-          animatedHeader: Text(
-            'Hello Flutter',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 24,
-              fontWeight: FontWeight.bold,
-              backgroundColor: Colors.blue,
-            ),
-            textAlign: TextAlign.center,
-          ),
+          drawer: const MyDrawer(),
+          alignment: DrawerAlignment.start,
+          showDrawerOpener: true,
+          drawerOpenerTopMargin: 8,
+          animatedHeader: drawerHeader,
           headerHeight: 50,
-          child: MyHomePage(),
+          child: const MyHomePage(),
         ),
       ),
     );