+++ example/lib/my_home_page.dart
@@ -9,6 +9,7 @@ class MyHomePage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
+        centerTitle: true,
         title: const Text('A flutter demo'),
       ),
       body: Center(