+++ lib/flutter_drawers.dart
@@ -1,3 +1,4 @@
 library flutter_custom_drawers;
 
 export 'src/box_drawer.dart';
+export 'src/slide_drawer.dart';