+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.3
+
+* New drawer widget added
+### SlideDrawer
+
+## 0.0.2
+
+* Bug fix
+
 ## 0.0.1
 
-* This is first release with only single drawer widget BoxDrawer 
+* This is first release with only single drawer widget
+### BoxDrawer